Split form engine and components configuration
