+++ src/apps/Demo.GovUk.Forms.AboutYou/StartupConfiguration.cs
@@ -3,6 +3,7 @@ using Demo.GovUk.Forms.AboutYou.Application.Services;
 using Demo.GovUk.Forms.AboutYou.Builders;
 using GovUk.Forms.Application.Factories;
 using GovUk.Forms.Application.Services;
+using GovUk.Forms.Components.Extensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -14,12 +15,21 @@ public class StartupConfiguration : IHostingStartup
 {
     public void Configure(IWebHostBuilder builder)
     {
-        builder.ConfigureServices(services =>
+        builder.ConfigureServices((context, services) =>
         {
             services.AddSingleton<IFormFactory, AboutYouFormFactory>();
             services.AddKeyedTransient<IFormPrePopulationService, TestFormPrePopulationService>(WebInfo.Root);
             YourDetailsFlowchart flowchartBuilder = new();
             flowchartBuilder.Construct(services);
+            
+            services.AddComponents(context.Configuration);
+            services.AddFormEngine(context.Configuration);
+        });
+        
+        builder.Configure(app =>
+        {
+            app.UseComponents();
+            app.UseFormEngine();
         });
     }
 }
