+++ src/apps/Demo.GovUk.Forms.Bankruptcy/StartupConfiguration.cs
@@ -1,6 +1,7 @@
 using Demo.GovUk.Forms.Bankruptcy.Application.Factories;
 using Demo.GovUk.Forms.Bankruptcy.Builders;
 using GovUk.Forms.Application.Factories;
+using GovUk.Forms.Components.Extensions;
 using GovUk.Forms.Components.Resolvers;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -13,7 +14,7 @@ public class StartupConfiguration : IHostingStartup
 {
     public void Configure(IWebHostBuilder builder)
     {
-        builder.ConfigureServices(services =>
+        builder.ConfigureServices((context, services) =>
         {
             services.AddSingleton<IFormFactory, BankruptcyFormFactory>();
             
@@ -21,6 +22,15 @@ public class StartupConfiguration : IHostingStartup
             flowchartBuilder.Construct(services);
 
             services.AddSingleton<IStartPageResolver, StartPageResolver>();
+            
+            services.AddComponents(context.Configuration);
+            services.AddFormEngine(context.Configuration);
+        });
+        
+        builder.Configure(app =>
+        {
+            app.UseComponents();
+            app.UseFormEngine();
         });
     }
 }
