+++ src/apps/Demo.GovUk.Forms.ContactUs/StartupConfiguration.cs
@@ -3,6 +3,7 @@ using Demo.GovUk.Forms.ContactUs.Builders;
 using Demo.GovUk.Forms.ContactUs.Infrastructure.Clients;
 using GovUk.Forms.Application.Extensions;
 using GovUk.Forms.Application.Factories;
+using GovUk.Forms.Components.Extensions;
 using GovUk.Forms.Infrastructure.Extensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -40,6 +41,15 @@ public class StartupConfiguration : IHostingStartup
             
             ContactUsFlowchart flowchartBuilder = new();
             flowchartBuilder.Construct(services);
+            
+            services.AddComponents(context.Configuration);
+            services.AddFormEngine(context.Configuration);
+        });
+        
+        builder.Configure(app =>
+        {
+            app.UseComponents();
+            app.UseFormEngine();
         });
     }
 }
