+++ src/apps/Inss.GovUk.Forms.Fip/WebInfo.cs
@@ -4,5 +4,5 @@ namespace Inss.GovUk.Forms.Fip;
 
 public static class WebInfo
 {
-    public static ContentPath Root => "/fip";
+    public static ContentPath Root => "/search";
 }
