+++ src/apps/Inss.GovUk.Forms.IPUpload/StartupConfiguration.cs
@@ -14,6 +14,7 @@ using System.Net.Http.Headers;
 using System.Net.Mime;
 using GovUk.Forms.Application.Factories;
 using GovUk.Forms.Components.Binding;
+using GovUk.Forms.Components.Extensions;
 using Inss.GovUk.Forms.IPUpload.Application.Factories;
 using Inss.GovUk.Forms.IPUpload.Domain;
 using Inss.GovUk.Forms.IPUpload.Domain.Validation;
@@ -73,6 +74,15 @@ public class StartupConfiguration : IHostingStartup
             
             IPUploadFlowchart flowchartBuilder = new();
             flowchartBuilder.Construct(services);
+            
+            services.AddComponents(context.Configuration);
+            services.AddFormEngine(context.Configuration);
+        });
+        
+        builder.Configure(app =>
+        {
+            app.UseComponents();
+            app.UseFormEngine();
         });
     }
 }
