--- src/engine/GovUk.Forms.Components/Options/ComponentOptions.cs
@@ -1,7 +1,0 @@
-﻿// ReSharper disable UnusedAutoPropertyAccessor.Global
-namespace GovUk.Forms.Components.Options;
-
-public sealed class ComponentOptions
-{
-    public bool BootstrapFormFramework { get; init; }
-}
