+++ src/host/GovUk.Forms.HostApp/appsettings.json
@@ -19,9 +19,6 @@
     "ClientSecret": "",
     "TenantId": ""
   },
-  "Components": {
-    "BootstrapFormFramework": true
-  },
   "Header": {
     "HomeLink": "https://localhost:5056",
     "Service": {
